pyproject.toml
[tool.poetry]
name = "skillhub_parsers"
version = "0.1.0"
description = ""
authors = ["Andrei Kashin <kashin.andrew@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.1"
loguru = "^0.5.3"
beautifulsoup4 = "^4.9.3"
lxml = "^4.6.3"
python-dotenv = "^0.18.0"
PySocks = "^1.7.1"
socks = "^0"
undetected-chromedriver = "^3.1.5"
selenium = "^4.1.3"
cloudscraper = "^1.2.60"
2captcha-python = "^1.1.2"
APScheduler = "^3.9.1"
yarl = "^1.8.1"
user-agent = "^0.1.10"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
